Training example of thickunet
Training of 200 thickslices + 18 of validatoin (1 patient) takes ~120s. After each epoch the model is saved and the statistics are logged in tensorboard. To make and actual train I would increase the number of slices per epoch to 1000-2000 and the number of validation patients to 10 hence each epoch would take 1200s which is reasonable.